+++ packages/tea-map-cache/tests/test-index.js
@@ -4,20 +4,13 @@ const test = require('ava');
 const proxyquire = require('proxyquire').noPreserveCache();
 
 test('tea-map-cache handler writes the expected bucketmap', async (t) => {
-  const { handler } = proxyquire('../index.js', {
+  const { handler } = proxyquire('../dist/index.js', {
     './tea': {
       getTeaBucketPath: async () => 'tea_bucket_path'
     },
-    'aws-sdk': {
-      S3: class mockedS3 {
-        putObject(params) {
-          return {
-            promise: async () => {
-              t.is(params.Body, '{"someBucket":"tea_bucket_path"}');
-            }
-          };
-        }
-      }
+    '@cumulus/aws-client/S3': {
+      s3PutObject:
+        async (params) => t.is(params.Body, '{"someBucket":"tea_bucket_path"}')
     }
   });
   process.env.TEA_API = 'https://foo/bar';